Fix tweets that lack information
Some of the PS4 tweets just contained hashtags. They are preserved
as tags on the post already, they don’t need to be in the body.

And some tweets happened in the same minute, so when creating a
generic tweet title it helps to include the seconds otherwise files
get overwritten.