Change the body font used in prose articles
I prefer a serif for reading a lot of copy. But I was using Lato for
that before. Partly because I've used it elsewhere and was just being
lazy, and partly as a contrast to Vollkorn for the main headers.

Vollkorn isn't a good body font. Well, at least, not to my eyes.
PT Serif I can read for ages though, so I'm using that. And where
I'm using a sans-serif (on the index pages, and misc other places)
I'm now using Assistant instead of Lato.