Add error details to the body of a Twist post
I didn't immediately notice the pull content failure this morning.
Despite there being a successful post to Twist, nothing shows in the Mac
app. The post does show in iOS but the theoretical title is ignored in
favour of the integration's title, so it is unclear what the post was.