Improve the logic of the Sectile dimensions
After a lot of testing Flourish v0.10's new Sectile fragments editing
interface in the blueprint page, I realised I hadn't been using the
source/type dimensions well.

"source" is an overloaded term, meaning both a page source (a TOML or
Markdown file definining a page), an output web page (Flourish calls the
generator for most sources SourceGenerator, and I am using `source` as
the reference name in `generate.py`) so using it as the dimension to
represent "the source of the source" was one to many.

This commit changes the dimension "source" to "origin" to be clearer,
but doesn't change any source to use the new dimension. That will
come in the next set of commits.