Restrict the number of CloudFront invalidations
When a common template changes, most (if not all) of the site's HTML
changes. Even when it is just different whitespace, and no actual
changed content.

As the HTML is not minified, this results in the entire site being
reuploaded to S3. And this triggers a mass CloudFront invalidation.
There are two problems caused by this:

1.  There's a limit of 3,000 invalidations in one request, and Flourish
    wasn't honouring that.
2.  Anything over 1,000 invalidations in a month is charged.

Invalidating the entire site would result in a (failed) request
containing approximately seven to nine *thousand* paths, charged at
around $35, insulting given it doesn't even do anything because of an
error.  Invalidating a wildcarded path of `/*` would do the same job of
invalidating the entire site, wouldn't fail, and would only cost $0.005.

Now when multiple pages are reuploaded and invalidated Flourish
will collapse common paths if necessary to bring the number of
invalidations down to no more than 50. Which should bring my AWS
bill down considerably.