Provide a quick build/push for changed posts
The downside to building the entire site before pushing it to S3 every
time is that flourish is not quick. It is taking around 15 minutes to
generate everything. So publishing a new post is not always timely.

By getting a list of which source files have changed and generating
only the pages that might be affected as a step before building the
entire site it should make new posts appear much quicker, and the
full rebuild may produce no other changes, but it is still there as
a safety net (and for when CSS or templates change and ALL pages
change as a result).